==> notebooks/1.0-extract-titanic-data.ipynb <==
#!pip install python-dotenv
! python -c 'python-dotenv' || pip python-dotenv --user

import os
import requests
from requests import session
from dotenv import load_dotenv, find_dotenv

#! pip install kaggle
! python -c 'kaggle' || pip install kaggle --user

download_dir = os.path.join (os.path.pardir, 'data', 'raw')

#!kaggle competitions download -c titanic --force -p ..\data\raw
!kaggle competitions download -c titanic -p ..\data\raw

file_name = !ls -lrt ../data/raw

print(file_name)

extract_path = os.path.join(download_dir, "titanic.zip")
print(extract_path)

from zipfile import ZipFile
zf = ZipFile(extract_path, 'r')
zf.extractall(download_dir)
zf.close()

train_data_path = os.path.join(download_dir, 'train.csv')
test_data_path= os.path.join(download_dir, 'test.csv')

print(train_data_path)
print(test_data_path)
