==> notebooks/3.0-building-predictive-models.ipynb <==
### Building Predictive Models - Part I

import pandas as pd
import os
import numpy as np

#### Import Data

# set the path of the processed data
processed_data_path = os.path.join(os.path.pardir, 'data', 'processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

train_df = pd.read_csv(train_file_path, index_col = 'PassengerId')
test_df = pd.read_csv(test_file_path, index_col = 'PassengerId')

train_df.info()

test_df.info()

#### Data Preparation

X = train_df.loc[:, 'Age':].values.astype('float') # Take all rows & columnns from starting from Age and convert their data type to real float
y = train_df['Survived'].ravel() # For output, use ravel() which is a numpy function which creates a flattend 1-D arrary

print(X.shape, y.shape)

# Perform the train test split using the training data
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 0) # random_state is the used as a seed value or randomization value
print(X_train.shape, y_train.shape)
print(X_test.shape, y_test.shape)                                                

# Average survival in train and test
print('Mean Survival in train data: {0:.3f}'.format(np.mean(y_train)))
print('Mean Survival in test data: {0:.3f}'.format(np.mean(y_test)))

##### Validate the Scikit-Learn Version

import sklearn

sklearn.__version__

#### Baseline the Model

# import function
from sklearn.dummy import DummyClassifier # To create a baseline model import DummyClassifier

# Creating a model
model_dummy = DummyClassifier(strategy = 'most_frequent', random_state = 0)

# Training a model
model_dummy.fit(X_train, y_train) # It takes input and output data

print('Score for the Baseline Model: {0:.2f}'.format(model_dummy.score(X_test, y_test))) ## Accuracy

# Performance Metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

# Accuracy Score
print('Accracy for the Baseline Model is : {0:.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test))))

# Confusion Matrix
print('Confusion matrix for the Baseline Model is : \n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test))))

import warnings
warnings.filterwarnings('ignore')

# Precision Score
print('Precision Score for the Baseline Model is : {0:.2f}'.format(precision_score(y_test, model_dummy.predict(X_test))))

# Recall Score
print('Recall Score for the Baseline Model is : {0:.2f}'.format(recall_score(y_test, model_dummy.predict(X_test))))

### First Kaggle Submission

# Converting to the Matrix
test_X = test_df.values.astype('float') # Converting the test data frame to 2-D Matrix and converting matrix to floating type

# Make Predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived' : predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
submission_data_file = os.path.join(submission_data_path, '01_dummy.csv')

df_submission.to_csv(submission_data_file, index=False)
# index = False to supress the additional column which will be created in the final output file with indexes

 def get_submission_file(model, filename):
        # converting to the matrix
        test_X = test_df.values.astype('float')
        # make predictions
        predictions = model.predict(test_X)
        # dataframe for submission
        df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived' : predictions})
        # generating a submission file
        submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
        submission_data_file = os.path.join(submission_data_path, filename)
        # writing to the submission file
        df_submission.to_csv(submission_data_file, index=False)

# get submission file
get_submission_file(model_dummy, '01_dummy.csv')

!kaggle competitions submit -c titanic -f ..\data\external\01_dummy.csv -m "01 Baseline Model"

#### Logisitical Regression Model

# import function
from sklearn.linear_model import LogisticRegression

# Create the Model
model_lr_1 = LogisticRegression(random_state = 0)

# Train the model
model_lr_1.fit(X_train, y_train)

# Evaluate the model
print('Score for the Logistic Regression Model Version 1 is : {0:.2f}'.format(model_lr_1.score(X_test, y_test))) ## Accuracy

# Evaluating the performance Matrices

# Accuracy
print('Accracy for the Logistic Regression Model Version 1 is : {0:.2f}'.format(accuracy_score(y_test, model_lr_1.predict(X_test))))

# Confusion Matrix
print('Confusion matrix for the Logistic Regression Model Version 1 is : \n {0}'.format(confusion_matrix(y_test, model_lr_1.predict(X_test))))

# Precision Score
print('Precision Score for the Logistic Regression Model Version 1 is : {0:.2f}'.format(precision_score(y_test, model_lr_1.predict(X_test))))

# Precision Score
print('Recall Score for the Logistic Regression Model Version 1 is : {0:.2f}'.format(recall_score(y_test, model_lr_1.predict(X_test))))

# Model Coefficients
model_lr_1.coef_ # model coef or model weights or model weights

### Second Kaggle Submission

# Use the get submission file
get_submission_file(model_lr_1, '02_lr.csv')

!kaggle competitions submit -c titanic -f ..\data\external\02_lr.csv -m "02 Second Submission Using Logistical Regression Model"

### Building Predictive Models - Part II

#### Hyperparameter Optimization

## Base Model
model_lr = LogisticRegression(random_state = 0)

from sklearn.model_selection import GridSearchCV # 

parameters = {'C' : [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty' : ['l1', 'l2']}

clf = GridSearchCV(model_lr, param_grid = parameters, cv = 3)
# C parameter controls the penality strength

clf.fit(X_train, y_train)

clf.best_params_

print('best score : {0:.2f}'.format(clf.best_score_))

# Evaluate the model
print('Score for the Logistic Regression Model Version 3 is : {0:.2f}'.format(clf.score(X_test, y_test)))

### Third and Final Kaggle Submission

get_submission_file(clf, '03_lr.csv')

!kaggle competitions submit -c titanic -f ..\data\external\03_lr.csv -m "03 Third Submission Using Logistical Regression Model"

#### Feature Normalization and Standardization 

from sklearn.preprocessing import MinMaxScaler, StandardScaler

#### Feature Normalization

scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(), X_train_scaled[:,0].max()

# normalize the test data
X_test_scaled = scaler.transform(X_test)

#### Feature Standardization

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

# Base Model
model_lr = LogisticRegression()
parameters = {'C' : [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty' : ['l1', 'l2']}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 3)
clf.fit(X_train_scaled, y_train)

clf.best_score_

print('Score for the Logistic Regression Model Scaled Version is : {0:.2f}'.format(clf.score(X_test_scaled, y_test)))

#### Model Persistence

# import pickle library
import pickle

# create the file paths
model_file_path = os.path.join(os.path.pardir, 'models', 'lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir, 'models', 'lr_scaler.pkl')

# open the files to write
model_file_pickle = open(model_file_path, 'wb') # wb is for the binary format
scaler_file_pickle = open(scaler_file_path, 'wb')

# persist the model and scaler
pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

# close the files
model_file_pickle.close()
scaler_file_pickle.close()

#### Load the Persisted File

# Open the files in read mode
model_file_pickle = open(model_file_path, 'rb')
scaler_file_pickle = open(scaler_file_path, 'rb')

# load files
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)

# close the files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

X_test_scaled = scaler_loaded.transform(X_test)
print('Score for the Persisted Logistic Regression Model is : {0:.2f}'.format(clf_loaded.score(X_test_scaled, y_test)))
